pyproject.toml
[project]
name = "docstribeproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.36.0",
    "langchain-text-splitters>=1.0.0",
    "numpy>=1.24",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=2.2.2",
]
